refactoring day06
