+++ 2019/src/lib.rs
@@ -5,3 +5,4 @@ pub mod day02;
 pub mod day03;
 pub mod day04;
 pub mod day05;
+pub mod day06;
