+++ 2019/tests/day06_test.rs
@@ -1,0 +1,14 @@
+#[cfg(test)]
+mod tests {
+    use aoc2019::day06;
+
+    #[test]
+    fn part1_works() {
+        assert_eq!(251208, day06::part1());
+    }
+
+    #[test]
+    fn part2_works() {
+        assert_eq!(397, day06::part2());
+    }
+}
